Add Copydepo manufacturer to configs and setup
This update includes the addition of Copydepo as a manufacturer to the existing list in the config files and setup file. Also, newly defined fields related to Copydepo in URLs, import paths, and views have been set in setup controller. This allows for future transactions or tasks involving the added manufacturer.