Add feature to link products to shipping methods
The shipping methods controller (szallitasimodController) has been updated to include a product (Termek) attribute, allowing to associate specific products with shipping methods. Corresponding changes were made in the entities and receiver controller files to support the new feature. It includes functions for getting and setting the product, and for removing the link when necessary.