Update condition to display additional tabs in partner form
An additional condition has been added to the logic that dictates when certain tabs are shown in the partner form. Now, apart from checking if the 'mptngy' configuration is not set, the code also checks if 'mpt' is not set. This affects tabs showing 'Termékkategória kedvezmények', 'Termék kedvezmények', 'Banki adatok', and others.